layout/base/tests/test_bug646757.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=646757
-->
<head>
  <title>Test for Bug 646757</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body style="margin:0" id="body">
<div style="height:20.3px; width:400px; background:pink" id="d1"></div>
<div style="height:20px; width:400px; background:yellow" id="d2"></div>
<div style="height:9.7px; width:400px;" id="space1"></div>
<div style="height:20.7px; width:400px; background:pink" id="d3"></div>
<div style="height:20px; width:400px; background:yellow" id="d4"></div>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=646757">Mozilla Bug 646757</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script class="testbody" type="text/javascript">
function testPoint(x, y, id) {
  is(document.elementFromPoint(x, y).id, id,
     "checking element at " + x + "," + y);
}

/** Test for Bug 646757 **/
testPoint(200, 20, "d1");
testPoint(200, 20.2, "d1");
testPoint(200, 20.4, "d2");
testPoint(200, 21, "d2");

testPoint(200, 70, "d3");
testPoint(200, 70.6, "d3");
testPoint(200, 70.8, "d4");
testPoint(200, 71, "d4");
</script>
</pre>
</body>
</html>
